test: split the background-download family into tests/test_downloads.py
tests/test_wrapper.py had grown to 6,635 lines and owned several whole tool
groups alongside the genuine wrapper core, against the AGENTS.md convention
that each tool group gets its own file.

Move the download_model / download_status / wait_for_download /
cancel_download family — argument guards, the submit-then-poll background
flow, the companions, and the legacy synchronous fallback — to
tests/test_downloads.py. Pure motion: no assertion, fixture or helper body
changed. _NO_SUCH_OPTION_STDERR stays in test_wrapper.py (the
_is_missing_option_error tests still use it) and is imported from there.

Refs #123
